+++ resources/views/livewire/attendance/attendance-display.blade.php
@@ -120,17 +120,8 @@
                             /** @var \App\Models\Member $member */
                             $attendance = $member->attendances()->where("event_id", $event->id)->first();
                             if($attendance === null) continue;
-                            /** @var \App\Models\Attendance $attendance */
-                            $cssClasses = $attendance?->attended ? " bg-green-300" : '';
                         @endphp
-                        <div class="flex gap-2 items-center px-2">
-                            <div class="h-2 w-2 rounded-full {{match($attendance?->poll_status){
-                        "in" => 'bg-green-700',
-                        "unsure" => 'bg-yellow-600',
-                        "out" => 'bg-red-700',
-                        default => ''} }}"></div>
-                            <span class="rounded px-2 {{$cssClasses}}">{{__($member->getFullName())}}</span>
-                        </div>
+                        <x-attendance.list-item-display :attendance="$attendance" :member="$member"/>
                     @empty
                         <span>{{__("Currently no attendance information.")}}</span>
                     @endforelse